Deprecate `verdi node tree` and related utilities
They will be removed in `aiida-core==2.0.0` which will allow to drop the
`ete3` dependency. This CLI command was they only code using this lib.
